% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_con.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Convert Celsius to Fahrenheit}
\usage{
celsius_to_fahrenheit(celsius)
}
\arguments{
\item{celsius}{A numeric vector of temperatures in Celsius.}
}
\value{
A numeric vector of temperatures converted to Fahrenheit.
}
\description{
This function converts temperatures from Celsius to Fahrenheit.
}
\examples{
celsius_to_fahrenheit(0)
celsius_to_fahrenheit(c(0, 25, 37))

}
