% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_con.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(fahrenheit)
}
\arguments{
\item{fahrenheit}{A numeric vector of temperatures in Fahrenheit.}
}
\value{
A numeric vector of temperatures converted to Celsius.
}
\description{
This function converts temperatures from Fahrenheit to Celsius.
}
\examples{
fahrenheit_to_celsius(32)
fahrenheit_to_celsius(c(32, 77, 98.6))

}
